Refactor types
Remove "global" `types.ts` file. Create `types` directory.
